allennlp/modules/image_encoders/convolution_2d.py
from typing import Tuple, Union

import torch

from allennlp.common.from_params import FromParams

class Convolution2d(torch.nn.Conv2d, FromParams):
    """
    A simple wrapper around ``torch.nn.Conv2d`` that lets us construct this ``from_params``.
    """
    def __init__(self,
                 in_channels: int,
                 out_channels: int,
                 kernel_size: Union[int, Tuple[int, int]],
                 stride: Union[int, Tuple[int, int]] = 1,
                 padding: Union[int, Tuple[int, int]] = 0,
                 dilation: Union[int, Tuple[int, int]] = 1,
                 groups: int = 1,
                 bias: bool = True) -> None:
        super().__init__(in_channels=in_channels,
                         out_channels=out_channels,
                         kernel_size=kernel_size,
                         stride=stride,
                         padding=padding,
                         dilation=dilation,
                         groups=groups,
                         bias=bias)
